requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.entry-points-selectable==1.1.0
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
black==22.3.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy==1.0.7
cryptography @ file:///C:/b/abs_8ecplyc3n2/croot/cryptography_1677533105000/work
cycler==0.11.0
Cython==0.29.33
-e git+https://github.com/samson-wang/cython_bbox.git@17a9209162d2107c1d2d17ea66fb2a3c7fa0c73e#egg=cython_bbox
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@96c752ce821a3340e27edd51c28a00665dd32a30
distlib==0.3.2
easydict==1.10
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fairscale==0.4.13
filelock==3.0.12
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
halpecocotools==0.0.0
huggingface-hub==0.13.1
hydra-core==1.3.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1672612343532/work
importlib-resources==5.12.0
install==1.3.5
iopath==0.1.9
ipykernel @ file:///D:/bld/ipykernel_1655369313836/work
ipython @ file:///D:/bld/ipython_1677617262387/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jsonpatch==1.32
jsonpointer==2.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1676579893731/work
jupyter_core @ file:///D:/bld/jupyter_core_1675110009198/work
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
munkres==1.1.4
mypy-extensions==1.0.0
natsort==8.3.1
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.0
numpy @ file:///C:/b/abs_datssh7cer/croot/numpy_and_numpy_base_1672336199388/work
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python-headless==4.7.0.72
opencv-python==4.7.0.72
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathspec==0.11.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs==2.3.0
portalocker==2.7.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.3
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1672682006896/work
pyOpenSSL @ file:///C:/b/abs_552w85x1jz/croot/pyopenssl_1677607703691/work
pyparsing==3.0.9
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pywin32==227
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657615952984/work
requests @ file:///C:/ci/requests_1657735342357/work
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
terminaltables==3.1.10
timm==0.6.12
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado @ file:///D:/bld/tornado_1656937934674/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
typing_extensions @ file:///C:/b/abs_89eui86zuq/croot/typing_extensions_1669923792806/work
urllib3 @ file:///C:/b/abs_9bcwxczrvm/croot/urllib3_1673575521331/work
virtualenv==20.7.2
visdom==0.2.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
websocket-client==1.5.1
Werkzeug==2.2.3
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
yacs==0.1.8
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
#!pip install -e git+https://github.com/samson-wang/cython_bbox.git#egg=cython-bbox
